26-FE: Move router service into its own folder
